Moved renaming tool
