+++ Arma3ObjectBuilder/tool_renaming/utils.py
