+++ Arma3ObjectBuilder/ui/__init__.py
@@ -11,7 +11,6 @@ from . import props_object_mesh
 from . import tool_outliner
 from . import tool_hitpoint
 from . import tool_mass
-from . import tool_paths
 from . import tool_proxies
 from . import tool_rigging
 from . import tool_scripts
