+++ Arma3ObjectBuilder/utilities/__init__.py
@@ -1,7 +1,6 @@
 # In order of dependency
 from . import clouds
 from . import proxy
-from . import renaming
 from . import masses
 from . import outliner
 from . import rigging
